[BE] Feat : subscribes 상세 구현 및 예외 처리

+++ backend/planty-service/src/main/java/com/planty/api/subscribe/service/SubscribeService.java
@@ -1,5 +1,6 @@
 package com.planty.api.subscribe.service;
 
+import com.planty.api.subscribe.request.UserSubscribeRequest;
 import com.planty.api.subscribe.response.UserSubscribeDatailResponse;
 import com.planty.api.subscribe.response.UserSubscribeResponse;
 
@@ -7,5 +8,7 @@ import java.util.List;
 
 public interface SubscribeService {
     List<UserSubscribeResponse> getUserSubscribe(String userId);
-    UserSubscribeDatailResponse getUserSubscribeDetail(String userId, Integer sid);
+    UserSubscribeDatailResponse getUserSubscribeDetail(String userId, Long sid);
+
+//    UserSubscribeRequest regSubscribe (String userId, Long spid);
 }
